requirements.txt
opencv-python
matplotlib
keras-ocr
numpy
pillow
tensorflow==2.15.0